requirements.txt
bs4
python-telegram-bot<14.0.0
requests
playwright
sentry-sdk
python-dotenv